Stripped implementation from .hpp file
